--- a/lib/collection/url.js
+++ b/lib/collection/url.js
@@ -105,7 +105,7 @@ _.assign(Url.prototype, /** @lends Url.prototype */ {
         if (_.isString(path)) {
             path && (path = path.replace(regexes.trimPath, MATCH_1)); // remove leading slash for valid path
             // if path is blank string, we set it to undefined, if '/' then single blank string array
-            path = !path ? undefined : (path === PATH_SEPARATOR ? [E] : path.split(PATH_SEPARATOR));
+            path = path ? (path === PATH_SEPARATOR ? [E] : path.split(PATH_SEPARATOR)) : undefined;
         }
 
         // expand host string
@@ -396,7 +396,7 @@ _.assign(Url, /** @lends Url */ {
             url = url.substr(p.path.length);
             p.path && (p.path = p.path.replace(regexes.trimPath, MATCH_1)); // remove leading slash for valid path
             // if path is blank string, we set it to undefined, if '/' then single blank string array
-            p.path = !p.path ? undefined : (p.path === PATH_SEPARATOR ? [E] : p.path.split(PATH_SEPARATOR));
+            p.path = p.path ? (p.path === PATH_SEPARATOR ? [E] : p.path.split(PATH_SEPARATOR)) : undefined;
         }
 
         // extract the query string