Fixes postmanlabs/postman-app-support#4663

- Handle empty path issue
- Properly parse string path in JSON representation