@kamalaknn 2018-07-17
@codenirvana This is updating a very critical code path. Can you do some benchmark tests?