@codenirvana 2018-07-17
@kamalaknn Similar performance.

**Before:**
```
1000 String Input   : 54.990ms
String + JSON Input : 1.237ms
Path Edge Case      : 0.229ms
Very Large Input    : 1.407ms

# 100 Iterations : 25.45s
```

**After:**
```
1000 String Input   : 54.742ms
String + JSON Input : 1.221ms
Path Edge Case      : 0.176ms
Very Large Input    : 1.370ms

# 100 Iterations : 25.50s
```